requirements.txt
numpy==1.22.1
keras==2.10.0
tensorflow==2.10.0
opencv-python==4.5.5.62
Pillow=9.3.0